config/requirements.txt
Django==3.1.6
djangorestframework==3.12.2
drf-yasg==1.20.0
gunicorn==20.0.4
django-cors-headers==3.7.0
whitenoise==5.2.0